+++ go.mod
@@ -5,9 +5,9 @@ go 1.17
 require (
 	github.com/antonfisher/nested-logrus-formatter v1.3.0
 	github.com/free5gc/logger_conf v1.0.0
-	github.com/free5gc/logger_util v1.0.0
+	github.com/free5gc/util v1.0.6
 	github.com/pkg/errors v0.9.1
-	github.com/sirupsen/logrus v1.7.0
+	github.com/sirupsen/logrus v1.8.1
 	github.com/stretchr/testify v1.8.3
 	github.com/vishvananda/netlink v1.1.0
 )
@@ -33,6 +33,7 @@ require (
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pelletier/go-toml/v2 v2.0.8 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/tim-ywliu/nested-logrus-formatter v1.3.2 // indirect
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
 	github.com/ugorji/go/codec v1.2.11 // indirect
 	github.com/vishvananda/netns v0.0.0-20191106174202-0a2b9b5464df // indirect
