--- internal/types/types.go
@@ -1,7 +1,0 @@
-package types
-
-// Interfaces
-type IKECrypto interface {
-	Encrypt(plainText []byte) ([]byte, error)
-	Decrypt(cipherText []byte) ([]byte, error)
-}
