+++ security/esn/esn.go
@@ -2,7 +2,6 @@ package esn
 
 import (
 	"github.com/pkg/errors"
-
 	"github.com/sirupsen/logrus"
 
 	"github.com/free5gc/ike/internal/logger"
