+++ security/integ/auth_hmac_sha2_256_128.go
