+++ security/prf/prf.go
@@ -4,7 +4,6 @@ import (
 	"hash"
 
 	"github.com/pkg/errors"
-
 	"github.com/sirupsen/logrus"
 
 	"github.com/free5gc/ike/internal/logger"
